[BUGFIX] Fix processor runner mixing up content object states
The processor runner is declared as a service, so it will only be
instantiated once. This caused issues with multiple nested calls to the
same processor runners which kept overriding their internal content
object renderer state.

Keep the content object renderer on the call stack to isolate it in the
multiple calls scenario.