[BUGFIX] Fix content object renderer states getting mixed up
The processor runner is declared as a service, so it will only be
instantiated once and shared between all dependents. Do not
keep the object renderer as a property, as multiple calls to the
respective inject method on the processor runner will cause a
mixup of state.

Keep the content object renderer isolated from each other on
the call stack.