[BUGFIX] Remove null content objects from compound processor response
Content element processors are allowed to return null to indicate that
their content element shall not be rendered at all. Respect the null
return values in the compound processor and remove the content element
from the compound processor response entirely. This way the content
element will not be passed to the view.