[BUGFIX] Fix rendering of USER_INT objects in the compound processor
The compound processor needs to reconfigure the format of extbase
content elements to 'json', so it can parse the results.

The old implementation had two drawbacks:
* The compound processor would only alter real content elements
tt_content.* and not tt_content.list elements.
* The content element would be forced to render json, even it was
rendered to replace <!--INT… markers in a cached page. This could lead
to json output on HTML pages.

Remove the code that was previously used to reconfigure the format of
extbase content elements. Allow the view resolver for extbase actions
to differentiate between cached and uncached rendering. This allows for
json output for the compound processor and html output if an uncached
objects gets replaced in the generated HTML.

Add a new option to the compound processor that allows to override the
uncached format. This is necessary if the output format of the entire
page is switched via TypoScript. In this case the output format for
uncached objects needs to be overridden to match the global output
format.