[BUGFIX] Fix rendering of USER_INT objects in the compound processor
The compound processor used to reconfigure the format of extbase
content elements to 'json', so it could parse the results. The
implementation had two drawbacks:

* The compound processor would only alter real content elements
tt_content.* and not tt_content.list elements.
* The content element would be forced to render json, even it was
rendered to replace <!--INT… markers in a cached page. This could lead
to json output on HTML pages.

Remove the code that was previously used to reconfigure the content
element. Now an ExtbaseViewResolver has to take care to switch between
json and html views itself. A sample for this will be added to the
pti_demo extension.

For extension that relied on the compound processor to switch extbase
to json, this is a breaking change.